@@ -1,18 +1,12 @@
 {
     "name": "telekurier/hermes-retresco-client",
     "require": {
-        "php": ">=7.0",
-        "symfony/serializer": "^3.1.0",
-        "symfony/property-access": "^3.1.0",
-        "guzzlehttp/guzzle": "^6.3.0",
-        "jane/jane": "^3.0.0",
-        "jane/open-api": "^3.0.0",
-        "jane/runtime": "^2.0",
-        "jane/openapi-runtime": "2.0"
+        "php": ">=7.1",
+        "jane-php/open-api-runtime": "^4.0"
     },
     "require-dev": {
-        "fr3d/swagger-assertions": "^0.8.3",
-        "phpunit/phpunit": "^6.4.1"
+        "jane-php/open-api": "^4.0",
+        "friendsofphp/php-cs-fixer": "^2.7.3"
     },
     "autoload": {
         "psr-4": {
@@ -25,17 +19,9 @@
     "scripts": {
         "outdated-direct": "composer show -o -D",
         "generate": [
-            "rm -rf generated/* && ./vendor/bin/jane-openapi generate swagger.json telekurier\\\\RetrescoClient generated"
+            "rm -rf generated/* && ./vendor/bin/jane-openapi generate --config-file=jane-openapi.php"
         ],
-        "clean-generated-code": "rm -rf generated/Resource generated/Normalizer",
-        "swagger-ui": [
-            "bower --version  >/dev/null 2>&1 || (echo 'This command requires bower.\n Follow the instructions of http://bower.io/#install-bower.'>&2 && exit 1)",
-            "bower install",
-            "cp -r bower_components/swagger-ui/dist swagger-ui",
-            "cp swagger.json swagger-ui/",
-            "cp swagger-ui.html swagger-ui/index.html",
-            "(xdg-open http://localhost:8123&); php -S localhost:8123 -t swagger-ui"
-        ]
+        "clean-generated-code": "rm -rf generated/Resource generated/Normalizer"
     },
     "extra": {
         "branch-alias": {
