@@ -1,8 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Normalizer;
 
-use Joli\Jane\Runtime\Reference;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
@@ -10,32 +17,30 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
 class ElasticSearchAggregationNormalizer implements DenormalizerInterface, NormalizerInterface, DenormalizerAwareInterface, NormalizerAwareInterface
 {
     use DenormalizerAwareTrait;
     use NormalizerAwareTrait;
+
     public function supportsDenormalization($data, $type, $format = null)
     {
-        if ($type !== 'telekurier\\RetrescoClient\\Model\\ElasticSearchAggregation') {
-            return false;
-        }
-        return true;
+        return $type === 'telekurier\\RetrescoClient\\Model\\ElasticSearchAggregation';
     }
+
     public function supportsNormalization($data, $format = null)
     {
-        if ($data instanceof \telekurier\RetrescoClient\Model\ElasticSearchAggregation) {
-            return true;
-        }
-        return false;
+        return $data instanceof \telekurier\RetrescoClient\Model\ElasticSearchAggregation;
     }
-    public function denormalize($data, $class, $format = null, array $context = array())
+
+    public function denormalize($data, $class, $format = null, array $context = [])
     {
         if (!is_object($data)) {
             throw new InvalidArgumentException();
         }
         $object = new \telekurier\RetrescoClient\Model\ElasticSearchAggregation();
         if (property_exists($data, 'buckets')) {
-            $values = array();
+            $values = [];
             foreach ($data->{'buckets'} as $value) {
                 $values[] = $value;
             }
@@ -47,13 +52,15 @@ class ElasticSearchAggregationNormalizer implements DenormalizerInterface, Norma
         if (property_exists($data, 'value')) {
             $object->setValue($data->{'value'});
         }
+
         return $object;
     }
-    public function normalize($object, $format = null, array $context = array())
+
+    public function normalize($object, $format = null, array $context = [])
     {
         $data = new \stdClass();
         if (null !== $object->getBuckets()) {
-            $values = array();
+            $values = [];
             foreach ($object->getBuckets() as $value) {
                 $values[] = $value;
             }
@@ -65,6 +72,7 @@ class ElasticSearchAggregationNormalizer implements DenormalizerInterface, Norma
         if (null !== $object->getValue()) {
             $data->{'value'} = $object->getValue();
         }
+
         return $data;
     }
 }
