@@ -1,8 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Normalizer;
 
-use Joli\Jane\Runtime\Reference;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
@@ -10,25 +17,23 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
 class ElasticSearchRawResultNormalizer implements DenormalizerInterface, NormalizerInterface, DenormalizerAwareInterface, NormalizerAwareInterface
 {
     use DenormalizerAwareTrait;
     use NormalizerAwareTrait;
+
     public function supportsDenormalization($data, $type, $format = null)
     {
-        if ($type !== 'telekurier\\RetrescoClient\\Model\\ElasticSearchRawResult') {
-            return false;
-        }
-        return true;
+        return $type === 'telekurier\\RetrescoClient\\Model\\ElasticSearchRawResult';
     }
+
     public function supportsNormalization($data, $format = null)
     {
-        if ($data instanceof \telekurier\RetrescoClient\Model\ElasticSearchRawResult) {
-            return true;
-        }
-        return false;
+        return $data instanceof \telekurier\RetrescoClient\Model\ElasticSearchRawResult;
     }
-    public function denormalize($data, $class, $format = null, array $context = array())
+
+    public function denormalize($data, $class, $format = null, array $context = [])
     {
         if (!is_object($data)) {
             throw new InvalidArgumentException();
@@ -41,7 +46,7 @@ class ElasticSearchRawResultNormalizer implements DenormalizerInterface, Normali
             $object->setTimedOut($data->{'timed_out'});
         }
         if (property_exists($data, 'aggregations')) {
-            $values = new \ArrayObject(array(), \ArrayObject::ARRAY_AS_PROPS);
+            $values = new \ArrayObject([], \ArrayObject::ARRAY_AS_PROPS);
             foreach ($data->{'aggregations'} as $key => $value) {
                 $values[$key] = $this->denormalizer->denormalize($value, 'telekurier\\RetrescoClient\\Model\\ElasticSearchAggregation', 'json', $context);
             }
@@ -50,9 +55,11 @@ class ElasticSearchRawResultNormalizer implements DenormalizerInterface, Normali
         if (property_exists($data, 'hits')) {
             $object->setHits($this->denormalizer->denormalize($data->{'hits'}, 'telekurier\\RetrescoClient\\Model\\ElasticSearchResult', 'json', $context));
         }
+
         return $object;
     }
-    public function normalize($object, $format = null, array $context = array())
+
+    public function normalize($object, $format = null, array $context = [])
     {
         $data = new \stdClass();
         if (null !== $object->getTook()) {
@@ -71,6 +78,7 @@ class ElasticSearchRawResultNormalizer implements DenormalizerInterface, Normali
         if (null !== $object->getHits()) {
             $data->{'hits'} = $this->normalizer->normalize($object->getHits(), 'json', $context);
         }
+
         return $data;
     }
 }
