@@ -1,8 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Normalizer;
 
-use Joli\Jane\Runtime\Reference;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
@@ -10,32 +17,30 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
 class ElasticSearchResultNormalizer implements DenormalizerInterface, NormalizerInterface, DenormalizerAwareInterface, NormalizerAwareInterface
 {
     use DenormalizerAwareTrait;
     use NormalizerAwareTrait;
+
     public function supportsDenormalization($data, $type, $format = null)
     {
-        if ($type !== 'telekurier\\RetrescoClient\\Model\\ElasticSearchResult') {
-            return false;
-        }
-        return true;
+        return $type === 'telekurier\\RetrescoClient\\Model\\ElasticSearchResult';
     }
+
     public function supportsNormalization($data, $format = null)
     {
-        if ($data instanceof \telekurier\RetrescoClient\Model\ElasticSearchResult) {
-            return true;
-        }
-        return false;
+        return $data instanceof \telekurier\RetrescoClient\Model\ElasticSearchResult;
     }
-    public function denormalize($data, $class, $format = null, array $context = array())
+
+    public function denormalize($data, $class, $format = null, array $context = [])
     {
         if (!is_object($data)) {
             throw new InvalidArgumentException();
         }
         $object = new \telekurier\RetrescoClient\Model\ElasticSearchResult();
         if (property_exists($data, 'hits')) {
-            $values = array();
+            $values = [];
             foreach ($data->{'hits'} as $value) {
                 $values[] = $this->denormalizer->denormalize($value, 'telekurier\\RetrescoClient\\Model\\RetrescoDocument', 'json', $context);
             }
@@ -47,13 +52,15 @@ class ElasticSearchResultNormalizer implements DenormalizerInterface, Normalizer
         if (property_exists($data, 'total')) {
             $object->setTotal($data->{'total'});
         }
+
         return $object;
     }
-    public function normalize($object, $format = null, array $context = array())
+
+    public function normalize($object, $format = null, array $context = [])
     {
         $data = new \stdClass();
         if (null !== $object->getHits()) {
-            $values = array();
+            $values = [];
             foreach ($object->getHits() as $value) {
                 $values[] = $this->normalizer->normalize($value, 'json', $context);
             }
@@ -65,6 +72,7 @@ class ElasticSearchResultNormalizer implements DenormalizerInterface, Normalizer
         if (null !== $object->getTotal()) {
             $data->{'total'} = $object->getTotal();
         }
+
         return $data;
     }
 }
