@@ -1,8 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Normalizer;
 
-use Joli\Jane\Runtime\Reference;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
@@ -10,25 +17,23 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
 class LocationNormalizer implements DenormalizerInterface, NormalizerInterface, DenormalizerAwareInterface, NormalizerAwareInterface
 {
     use DenormalizerAwareTrait;
     use NormalizerAwareTrait;
+
     public function supportsDenormalization($data, $type, $format = null)
     {
-        if ($type !== 'telekurier\\RetrescoClient\\Model\\Location') {
-            return false;
-        }
-        return true;
+        return $type === 'telekurier\\RetrescoClient\\Model\\Location';
     }
+
     public function supportsNormalization($data, $format = null)
     {
-        if ($data instanceof \telekurier\RetrescoClient\Model\Location) {
-            return true;
-        }
-        return false;
+        return $data instanceof \telekurier\RetrescoClient\Model\Location;
     }
-    public function denormalize($data, $class, $format = null, array $context = array())
+
+    public function denormalize($data, $class, $format = null, array $context = [])
     {
         if (!is_object($data)) {
             throw new InvalidArgumentException();
@@ -40,9 +45,11 @@ class LocationNormalizer implements DenormalizerInterface, NormalizerInterface, 
         if (property_exists($data, 'lon')) {
             $object->setLon($data->{'lon'});
         }
+
         return $object;
     }
-    public function normalize($object, $format = null, array $context = array())
+
+    public function normalize($object, $format = null, array $context = [])
     {
         $data = new \stdClass();
         if (null !== $object->getLat()) {
@@ -51,6 +58,7 @@ class LocationNormalizer implements DenormalizerInterface, NormalizerInterface, 
         if (null !== $object->getLon()) {
             $data->{'lon'} = $object->getLon();
         }
+
         return $data;
     }
 }
