@@ -1,12 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Normalizer;
 
 class NormalizerFactory
 {
     public static function create()
     {
-        $normalizers = array();
+        $normalizers = [];
         $normalizers[] = new \Symfony\Component\Serializer\Normalizer\ArrayDenormalizer();
         $normalizers[] = new LocationNormalizer();
         $normalizers[] = new PinNormalizer();
@@ -20,6 +28,7 @@ class NormalizerFactory
         $normalizers[] = new ElasticSearchRawResultNormalizer();
         $normalizers[] = new ElasticSearchAggregationNormalizer();
         $normalizers[] = new ElasticSearchResultNormalizer();
+
         return $normalizers;
     }
 }
