@@ -1,8 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Normalizer;
 
-use Joli\Jane\Runtime\Reference;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
@@ -10,25 +17,23 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
 class RetrescoDocumentNormalizer implements DenormalizerInterface, NormalizerInterface, DenormalizerAwareInterface, NormalizerAwareInterface
 {
     use DenormalizerAwareTrait;
     use NormalizerAwareTrait;
+
     public function supportsDenormalization($data, $type, $format = null)
     {
-        if ($type !== 'telekurier\\RetrescoClient\\Model\\RetrescoDocument') {
-            return false;
-        }
-        return true;
+        return $type === 'telekurier\\RetrescoClient\\Model\\RetrescoDocument';
     }
+
     public function supportsNormalization($data, $format = null)
     {
-        if ($data instanceof \telekurier\RetrescoClient\Model\RetrescoDocument) {
-            return true;
-        }
-        return false;
+        return $data instanceof \telekurier\RetrescoClient\Model\RetrescoDocument;
     }
-    public function denormalize($data, $class, $format = null, array $context = array())
+
+    public function denormalize($data, $class, $format = null, array $context = [])
     {
         if (!is_object($data)) {
             throw new InvalidArgumentException();
@@ -128,49 +133,49 @@ class RetrescoDocumentNormalizer implements DenormalizerInterface, NormalizerInt
             $object->setPin($this->denormalizer->denormalize($data->{'pin'}, 'telekurier\\RetrescoClient\\Model\\Pin', 'json', $context));
         }
         if (property_exists($data, 'rtr_persons')) {
-            $values = array();
+            $values = [];
             foreach ($data->{'rtr_persons'} as $value) {
                 $values[] = $value;
             }
             $object->setRtrPersons($values);
         }
         if (property_exists($data, 'rtr_locations')) {
-            $values_1 = array();
+            $values_1 = [];
             foreach ($data->{'rtr_locations'} as $value_1) {
                 $values_1[] = $value_1;
             }
             $object->setRtrLocations($values_1);
         }
         if (property_exists($data, 'rtr_organisations')) {
-            $values_2 = array();
+            $values_2 = [];
             foreach ($data->{'rtr_organisations'} as $value_2) {
                 $values_2[] = $value_2;
             }
             $object->setRtrOrganisations($values_2);
         }
         if (property_exists($data, 'rtr_products')) {
-            $values_3 = array();
+            $values_3 = [];
             foreach ($data->{'rtr_products'} as $value_3) {
                 $values_3[] = $value_3;
             }
             $object->setRtrProducts($values_3);
         }
         if (property_exists($data, 'rtr_events')) {
-            $values_4 = array();
+            $values_4 = [];
             foreach ($data->{'rtr_events'} as $value_4) {
                 $values_4[] = $value_4;
             }
             $object->setRtrEvents($values_4);
         }
         if (property_exists($data, 'rtr_keywords')) {
-            $values_5 = array();
+            $values_5 = [];
             foreach ($data->{'rtr_keywords'} as $value_5) {
                 $values_5[] = $value_5;
             }
             $object->setRtrKeywords($values_5);
         }
         if (property_exists($data, 'rtr_tags')) {
-            $values_6 = array();
+            $values_6 = [];
             foreach ($data->{'rtr_tags'} as $value_6) {
                 $values_6[] = $value_6;
             }
@@ -179,9 +184,11 @@ class RetrescoDocumentNormalizer implements DenormalizerInterface, NormalizerInt
         if (property_exists($data, 'payload')) {
             $object->setPayload($data->{'payload'});
         }
+
         return $object;
     }
-    public function normalize($object, $format = null, array $context = array())
+
+    public function normalize($object, $format = null, array $context = [])
     {
         $data = new \stdClass();
         if (null !== $object->getDocId()) {
@@ -278,49 +285,49 @@ class RetrescoDocumentNormalizer implements DenormalizerInterface, NormalizerInt
             $data->{'pin'} = $this->normalizer->normalize($object->getPin(), 'json', $context);
         }
         if (null !== $object->getRtrPersons()) {
-            $values = array();
+            $values = [];
             foreach ($object->getRtrPersons() as $value) {
                 $values[] = $value;
             }
             $data->{'rtr_persons'} = $values;
         }
         if (null !== $object->getRtrLocations()) {
-            $values_1 = array();
+            $values_1 = [];
             foreach ($object->getRtrLocations() as $value_1) {
                 $values_1[] = $value_1;
             }
             $data->{'rtr_locations'} = $values_1;
         }
         if (null !== $object->getRtrOrganisations()) {
-            $values_2 = array();
+            $values_2 = [];
             foreach ($object->getRtrOrganisations() as $value_2) {
                 $values_2[] = $value_2;
             }
             $data->{'rtr_organisations'} = $values_2;
         }
         if (null !== $object->getRtrProducts()) {
-            $values_3 = array();
+            $values_3 = [];
             foreach ($object->getRtrProducts() as $value_3) {
                 $values_3[] = $value_3;
             }
             $data->{'rtr_products'} = $values_3;
         }
         if (null !== $object->getRtrEvents()) {
-            $values_4 = array();
+            $values_4 = [];
             foreach ($object->getRtrEvents() as $value_4) {
                 $values_4[] = $value_4;
             }
             $data->{'rtr_events'} = $values_4;
         }
         if (null !== $object->getRtrKeywords()) {
-            $values_5 = array();
+            $values_5 = [];
             foreach ($object->getRtrKeywords() as $value_5) {
                 $values_5[] = $value_5;
             }
             $data->{'rtr_keywords'} = $values_5;
         }
         if (null !== $object->getRtrTags()) {
-            $values_6 = array();
+            $values_6 = [];
             foreach ($object->getRtrTags() as $value_6) {
                 $values_6[] = $value_6;
             }
@@ -329,6 +336,7 @@ class RetrescoDocumentNormalizer implements DenormalizerInterface, NormalizerInt
         if (null !== $object->getPayload()) {
             $data->{'payload'} = $object->getPayload();
         }
+
         return $data;
     }
 }
