@@ -1,8 +1,15 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Normalizer;
 
-use Joli\Jane\Runtime\Reference;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerAwareTrait;
@@ -10,32 +17,30 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerAwareTrait;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
+
 class RetrescoTopicPagesNormalizer implements DenormalizerInterface, NormalizerInterface, DenormalizerAwareInterface, NormalizerAwareInterface
 {
     use DenormalizerAwareTrait;
     use NormalizerAwareTrait;
+
     public function supportsDenormalization($data, $type, $format = null)
     {
-        if ($type !== 'telekurier\\RetrescoClient\\Model\\RetrescoTopicPages') {
-            return false;
-        }
-        return true;
+        return $type === 'telekurier\\RetrescoClient\\Model\\RetrescoTopicPages';
     }
+
     public function supportsNormalization($data, $format = null)
     {
-        if ($data instanceof \telekurier\RetrescoClient\Model\RetrescoTopicPages) {
-            return true;
-        }
-        return false;
+        return $data instanceof \telekurier\RetrescoClient\Model\RetrescoTopicPages;
     }
-    public function denormalize($data, $class, $format = null, array $context = array())
+
+    public function denormalize($data, $class, $format = null, array $context = [])
     {
         if (!is_object($data)) {
             throw new InvalidArgumentException();
         }
         $object = new \telekurier\RetrescoClient\Model\RetrescoTopicPages();
         if (property_exists($data, 'docs')) {
-            $values = array();
+            $values = [];
             foreach ($data->{'docs'} as $value) {
                 $values[] = $this->denormalizer->denormalize($value, 'telekurier\\RetrescoClient\\Model\\RetrescoTopicPage', 'json', $context);
             }
@@ -44,13 +49,15 @@ class RetrescoTopicPagesNormalizer implements DenormalizerInterface, NormalizerI
         if (property_exists($data, 'num_found')) {
             $object->setNumFound($data->{'num_found'});
         }
+
         return $object;
     }
-    public function normalize($object, $format = null, array $context = array())
+
+    public function normalize($object, $format = null, array $context = [])
     {
         $data = new \stdClass();
         if (null !== $object->getDocs()) {
-            $values = array();
+            $values = [];
             foreach ($object->getDocs() as $value) {
                 $values[] = $this->normalizer->normalize($value, 'json', $context);
             }
@@ -59,6 +66,7 @@ class RetrescoTopicPagesNormalizer implements DenormalizerInterface, NormalizerI
         if (null !== $object->getNumFound()) {
             $data->{'num_found'} = $object->getNumFound();
         }
+
         return $data;
     }
 }
