@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class ElasticSearchAggregation
@@ -16,40 +24,47 @@ class ElasticSearchAggregation
      * @var mixed
      */
     protected $value;
+
     /**
      * @return mixed[]
      */
-    public function getBuckets()
+    public function getBuckets(): ?array
     {
         return $this->buckets;
     }
+
     /**
      * @param mixed[] $buckets
      *
      * @return self
      */
-    public function setBuckets(array $buckets = null)
+    public function setBuckets(?array $buckets): self
     {
         $this->buckets = $buckets;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getDocCount()
+    public function getDocCount(): ?int
     {
         return $this->docCount;
     }
+
     /**
      * @param int $docCount
      *
      * @return self
      */
-    public function setDocCount($docCount = null)
+    public function setDocCount(?int $docCount): self
     {
         $this->docCount = $docCount;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -57,14 +72,16 @@ class ElasticSearchAggregation
     {
         return $this->value;
     }
+
     /**
      * @param mixed $value
      *
      * @return self
      */
-    public function setValue($value = null)
+    public function setValue($value): self
     {
         $this->value = $value;
+
         return $this;
     }
 }
