@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class ElasticSearchRawResult
@@ -20,23 +28,27 @@ class ElasticSearchRawResult
      * @var ElasticSearchResult
      */
     protected $hits;
+
     /**
      * @return int
      */
-    public function getTook()
+    public function getTook(): ?int
     {
         return $this->took;
     }
+
     /**
      * @param int $took
      *
      * @return self
      */
-    public function setTook($took = null)
+    public function setTook(?int $took): self
     {
         $this->took = $took;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -44,48 +56,56 @@ class ElasticSearchRawResult
     {
         return $this->timedOut;
     }
+
     /**
      * @param mixed $timedOut
      *
      * @return self
      */
-    public function setTimedOut($timedOut = null)
+    public function setTimedOut($timedOut): self
     {
         $this->timedOut = $timedOut;
+
         return $this;
     }
+
     /**
      * @return ElasticSearchAggregation[]
      */
-    public function getAggregations()
+    public function getAggregations(): ?\ArrayObject
     {
         return $this->aggregations;
     }
+
     /**
      * @param ElasticSearchAggregation[] $aggregations
      *
      * @return self
      */
-    public function setAggregations(\ArrayObject $aggregations = null)
+    public function setAggregations(?\ArrayObject $aggregations): self
     {
         $this->aggregations = $aggregations;
+
         return $this;
     }
+
     /**
      * @return ElasticSearchResult
      */
-    public function getHits()
+    public function getHits(): ?ElasticSearchResult
     {
         return $this->hits;
     }
+
     /**
      * @param ElasticSearchResult $hits
      *
      * @return self
      */
-    public function setHits(ElasticSearchResult $hits = null)
+    public function setHits(?ElasticSearchResult $hits): self
     {
         $this->hits = $hits;
+
         return $this;
     }
 }
