@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class ElasticSearchResult
@@ -16,55 +24,64 @@ class ElasticSearchResult
      * @var int
      */
     protected $total;
+
     /**
      * @return RetrescoDocument[]
      */
-    public function getHits()
+    public function getHits(): ?array
     {
         return $this->hits;
     }
+
     /**
      * @param RetrescoDocument[] $hits
      *
      * @return self
      */
-    public function setHits(array $hits = null)
+    public function setHits(?array $hits): self
     {
         $this->hits = $hits;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getMaxScore()
+    public function getMaxScore(): ?int
     {
         return $this->maxScore;
     }
+
     /**
      * @param int $maxScore
      *
      * @return self
      */
-    public function setMaxScore($maxScore = null)
+    public function setMaxScore(?int $maxScore): self
     {
         $this->maxScore = $maxScore;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getTotal()
+    public function getTotal(): ?int
     {
         return $this->total;
     }
+
     /**
      * @param int $total
      *
      * @return self
      */
-    public function setTotal($total = null)
+    public function setTotal(?int $total): self
     {
         $this->total = $total;
+
         return $this;
     }
 }
