@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class Location
@@ -12,6 +20,7 @@ class Location
      * @var mixed
      */
     protected $lon;
+
     /**
      * @return mixed
      */
@@ -19,16 +28,19 @@ class Location
     {
         return $this->lat;
     }
+
     /**
      * @param mixed $lat
      *
      * @return self
      */
-    public function setLat($lat = null)
+    public function setLat($lat): self
     {
         $this->lat = $lat;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -36,14 +48,16 @@ class Location
     {
         return $this->lon;
     }
+
     /**
      * @param mixed $lon
      *
      * @return self
      */
-    public function setLon($lon = null)
+    public function setLon($lon): self
     {
         $this->lon = $lon;
+
         return $this;
     }
 }
