@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class Pin
@@ -8,21 +16,24 @@ class Pin
      * @var Location
      */
     protected $location;
+
     /**
      * @return Location
      */
-    public function getLocation()
+    public function getLocation(): ?Location
     {
         return $this->location;
     }
+
     /**
      * @param Location $location
      *
      * @return self
      */
-    public function setLocation(Location $location = null)
+    public function setLocation(?Location $location): self
     {
         $this->location = $location;
+
         return $this;
     }
 }
