@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class RelatedDocuments
@@ -8,21 +16,24 @@ class RelatedDocuments
      * @var RetrescoDocument[]
      */
     protected $docs;
+
     /**
      * @return RetrescoDocument[]
      */
-    public function getDocs()
+    public function getDocs(): ?array
     {
         return $this->docs;
     }
+
     /**
      * @param RetrescoDocument[] $docs
      *
      * @return self
      */
-    public function setDocs(array $docs = null)
+    public function setDocs(?array $docs): self
     {
         $this->docs = $docs;
+
         return $this;
     }
 }
