@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class RetrescoClientError
@@ -12,38 +20,44 @@ class RetrescoClientError
      * @var string
      */
     protected $message;
+
     /**
      * @return int
      */
-    public function getStatus()
+    public function getStatus(): ?int
     {
         return $this->status;
     }
+
     /**
      * @param int $status
      *
      * @return self
      */
-    public function setStatus($status = null)
+    public function setStatus(?int $status): self
     {
         $this->status = $status;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getMessage()
+    public function getMessage(): ?string
     {
         return $this->message;
     }
+
     /**
      * @param string $message
      *
      * @return self
      */
-    public function setMessage($message = null)
+    public function setMessage(?string $message): self
     {
         $this->message = $message;
+
         return $this;
     }
 }
