@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class RetrescoDocument
@@ -160,227 +168,267 @@ class RetrescoDocument
      * @var mixed
      */
     protected $payload;
+
     /**
      * @return string
      */
-    public function getDocId()
+    public function getDocId(): ?string
     {
         return $this->docId;
     }
+
     /**
      * @param string $docId
      *
      * @return self
      */
-    public function setDocId($docId = null)
+    public function setDocId(?string $docId): self
     {
         $this->docId = $docId;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getDocType()
+    public function getDocType(): ?string
     {
         return $this->docType;
     }
+
     /**
      * @param string $docType
      *
      * @return self
      */
-    public function setDocType($docType = null)
+    public function setDocType(?string $docType): self
     {
         $this->docType = $docType;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getUrl()
+    public function getUrl(): ?string
     {
         return $this->url;
     }
+
     /**
      * @param string $url
      *
      * @return self
      */
-    public function setUrl($url = null)
+    public function setUrl(?string $url): self
     {
         $this->url = $url;
+
         return $this;
     }
+
     /**
      * @return bool
      */
-    public function getPublished()
+    public function getPublished(): ?bool
     {
         return $this->published;
     }
+
     /**
      * @param bool $published
      *
      * @return self
      */
-    public function setPublished($published = null)
+    public function setPublished(?bool $published): self
     {
         $this->published = $published;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getLifecycle()
+    public function getLifecycle(): ?string
     {
         return $this->lifecycle;
     }
+
     /**
      * @param string $lifecycle
      *
      * @return self
      */
-    public function setLifecycle($lifecycle = null)
+    public function setLifecycle(?string $lifecycle): self
     {
         $this->lifecycle = $lifecycle;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTitle()
+    public function getTitle(): ?string
     {
         return $this->title;
     }
+
     /**
      * @param string $title
      *
      * @return self
      */
-    public function setTitle($title = null)
+    public function setTitle(?string $title): self
     {
         $this->title = $title;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getSupertitle()
+    public function getSupertitle(): ?string
     {
         return $this->supertitle;
     }
+
     /**
      * @param string $supertitle
      *
      * @return self
      */
-    public function setSupertitle($supertitle = null)
+    public function setSupertitle(?string $supertitle): self
     {
         $this->supertitle = $supertitle;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getAuthor()
+    public function getAuthor(): ?string
     {
         return $this->author;
     }
+
     /**
      * @param string $author
      *
      * @return self
      */
-    public function setAuthor($author = null)
+    public function setAuthor(?string $author): self
     {
         $this->author = $author;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTeaser()
+    public function getTeaser(): ?string
     {
         return $this->teaser;
     }
+
     /**
      * @param string $teaser
      *
      * @return self
      */
-    public function setTeaser($teaser = null)
+    public function setTeaser(?string $teaser): self
     {
         $this->teaser = $teaser;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTeaserImgUrl()
+    public function getTeaserImgUrl(): ?string
     {
         return $this->teaserImgUrl;
     }
+
     /**
      * @param string $teaserImgUrl
      *
      * @return self
      */
-    public function setTeaserImgUrl($teaserImgUrl = null)
+    public function setTeaserImgUrl(?string $teaserImgUrl): self
     {
         $this->teaserImgUrl = $teaserImgUrl;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTeaserImgSubline()
+    public function getTeaserImgSubline(): ?string
     {
         return $this->teaserImgSubline;
     }
+
     /**
      * @param string $teaserImgSubline
      *
      * @return self
      */
-    public function setTeaserImgSubline($teaserImgSubline = null)
+    public function setTeaserImgSubline(?string $teaserImgSubline): self
     {
         $this->teaserImgSubline = $teaserImgSubline;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getBody()
+    public function getBody(): ?string
     {
         return $this->body;
     }
+
     /**
      * @param string $body
      *
      * @return self
      */
-    public function setBody($body = null)
+    public function setBody(?string $body): self
     {
         $this->body = $body;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getSection()
+    public function getSection(): ?string
     {
         return $this->section;
     }
+
     /**
      * @param string $section
      *
      * @return self
      */
-    public function setSection($section = null)
+    public function setSection(?string $section): self
     {
         $this->section = $section;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -388,16 +436,19 @@ class RetrescoDocument
     {
         return $this->date;
     }
+
     /**
      * @param mixed $date
      *
      * @return self
      */
-    public function setDate($date = null)
+    public function setDate($date): self
     {
         $this->date = $date;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -405,16 +456,19 @@ class RetrescoDocument
     {
         return $this->startdate;
     }
+
     /**
      * @param mixed $startdate
      *
      * @return self
      */
-    public function setStartdate($startdate = null)
+    public function setStartdate($startdate): self
     {
         $this->startdate = $startdate;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -422,16 +476,19 @@ class RetrescoDocument
     {
         return $this->enddate;
     }
+
     /**
      * @param mixed $enddate
      *
      * @return self
      */
-    public function setEnddate($enddate = null)
+    public function setEnddate($enddate): self
     {
         $this->enddate = $enddate;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -439,16 +496,19 @@ class RetrescoDocument
     {
         return $this->publishedDate;
     }
+
     /**
      * @param mixed $publishedDate
      *
      * @return self
      */
-    public function setPublishedDate($publishedDate = null)
+    public function setPublishedDate($publishedDate): self
     {
         $this->publishedDate = $publishedDate;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -456,101 +516,119 @@ class RetrescoDocument
     {
         return $this->updatedDate;
     }
+
     /**
      * @param mixed $updatedDate
      *
      * @return self
      */
-    public function setUpdatedDate($updatedDate = null)
+    public function setUpdatedDate($updatedDate): self
     {
         $this->updatedDate = $updatedDate;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getSource()
+    public function getSource(): ?string
     {
         return $this->source;
     }
+
     /**
      * @param string $source
      *
      * @return self
      */
-    public function setSource($source = null)
+    public function setSource(?string $source): self
     {
         $this->source = $source;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getAgentur()
+    public function getAgentur(): ?string
     {
         return $this->agentur;
     }
+
     /**
      * @param string $agentur
      *
      * @return self
      */
-    public function setAgentur($agentur = null)
+    public function setAgentur(?string $agentur): self
     {
         $this->agentur = $agentur;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getChannel()
+    public function getChannel(): ?string
     {
         return $this->channel;
     }
+
     /**
      * @param string $channel
      *
      * @return self
      */
-    public function setChannel($channel = null)
+    public function setChannel(?string $channel): self
     {
         $this->channel = $channel;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getArticleScore()
+    public function getArticleScore(): ?int
     {
         return $this->articleScore;
     }
+
     /**
      * @param int $articleScore
      *
      * @return self
      */
-    public function setArticleScore($articleScore = null)
+    public function setArticleScore(?int $articleScore): self
     {
         $this->articleScore = $articleScore;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getPiLast72h()
+    public function getPiLast72h(): ?int
     {
         return $this->piLast72h;
     }
+
     /**
      * @param int $piLast72h
      *
      * @return self
      */
-    public function setPiLast72h($piLast72h = null)
+    public function setPiLast72h(?int $piLast72h): self
     {
         $this->piLast72h = $piLast72h;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -558,254 +636,299 @@ class RetrescoDocument
     {
         return $this->gaLastSeen;
     }
+
     /**
      * @param mixed $gaLastSeen
      *
      * @return self
      */
-    public function setGaLastSeen($gaLastSeen = null)
+    public function setGaLastSeen($gaLastSeen): self
     {
         $this->gaLastSeen = $gaLastSeen;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getCommentsCount()
+    public function getCommentsCount(): ?int
     {
         return $this->commentsCount;
     }
+
     /**
      * @param int $commentsCount
      *
      * @return self
      */
-    public function setCommentsCount($commentsCount = null)
+    public function setCommentsCount(?int $commentsCount): self
     {
         $this->commentsCount = $commentsCount;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getSocialmediaShares()
+    public function getSocialmediaShares(): ?int
     {
         return $this->socialmediaShares;
     }
+
     /**
      * @param int $socialmediaShares
      *
      * @return self
      */
-    public function setSocialmediaShares($socialmediaShares = null)
+    public function setSocialmediaShares(?int $socialmediaShares): self
     {
         $this->socialmediaShares = $socialmediaShares;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getSocialmediaTraffic()
+    public function getSocialmediaTraffic(): ?int
     {
         return $this->socialmediaTraffic;
     }
+
     /**
      * @param int $socialmediaTraffic
      *
      * @return self
      */
-    public function setSocialmediaTraffic($socialmediaTraffic = null)
+    public function setSocialmediaTraffic(?int $socialmediaTraffic): self
     {
         $this->socialmediaTraffic = $socialmediaTraffic;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getRetentionPeriod()
+    public function getRetentionPeriod(): ?int
     {
         return $this->retentionPeriod;
     }
+
     /**
      * @param int $retentionPeriod
      *
      * @return self
      */
-    public function setRetentionPeriod($retentionPeriod = null)
+    public function setRetentionPeriod(?int $retentionPeriod): self
     {
         $this->retentionPeriod = $retentionPeriod;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getVideoViews()
+    public function getVideoViews(): ?int
     {
         return $this->videoViews;
     }
+
     /**
      * @param int $videoViews
      *
      * @return self
      */
-    public function setVideoViews($videoViews = null)
+    public function setVideoViews(?int $videoViews): self
     {
         $this->videoViews = $videoViews;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getBounceRate()
+    public function getBounceRate(): ?int
     {
         return $this->bounceRate;
     }
+
     /**
      * @param int $bounceRate
      *
      * @return self
      */
-    public function setBounceRate($bounceRate = null)
+    public function setBounceRate(?int $bounceRate): self
     {
         $this->bounceRate = $bounceRate;
+
         return $this;
     }
+
     /**
      * @return Pin
      */
-    public function getPin()
+    public function getPin(): ?Pin
     {
         return $this->pin;
     }
+
     /**
      * @param Pin $pin
      *
      * @return self
      */
-    public function setPin(Pin $pin = null)
+    public function setPin(?Pin $pin): self
     {
         $this->pin = $pin;
+
         return $this;
     }
+
     /**
      * @return string[]
      */
-    public function getRtrPersons()
+    public function getRtrPersons(): ?array
     {
         return $this->rtrPersons;
     }
+
     /**
      * @param string[] $rtrPersons
      *
      * @return self
      */
-    public function setRtrPersons(array $rtrPersons = null)
+    public function setRtrPersons(?array $rtrPersons): self
     {
         $this->rtrPersons = $rtrPersons;
+
         return $this;
     }
+
     /**
      * @return string[]
      */
-    public function getRtrLocations()
+    public function getRtrLocations(): ?array
     {
         return $this->rtrLocations;
     }
+
     /**
      * @param string[] $rtrLocations
      *
      * @return self
      */
-    public function setRtrLocations(array $rtrLocations = null)
+    public function setRtrLocations(?array $rtrLocations): self
     {
         $this->rtrLocations = $rtrLocations;
+
         return $this;
     }
+
     /**
      * @return string[]
      */
-    public function getRtrOrganisations()
+    public function getRtrOrganisations(): ?array
     {
         return $this->rtrOrganisations;
     }
+
     /**
      * @param string[] $rtrOrganisations
      *
      * @return self
      */
-    public function setRtrOrganisations(array $rtrOrganisations = null)
+    public function setRtrOrganisations(?array $rtrOrganisations): self
     {
         $this->rtrOrganisations = $rtrOrganisations;
+
         return $this;
     }
+
     /**
      * @return string[]
      */
-    public function getRtrProducts()
+    public function getRtrProducts(): ?array
     {
         return $this->rtrProducts;
     }
+
     /**
      * @param string[] $rtrProducts
      *
      * @return self
      */
-    public function setRtrProducts(array $rtrProducts = null)
+    public function setRtrProducts(?array $rtrProducts): self
     {
         $this->rtrProducts = $rtrProducts;
+
         return $this;
     }
+
     /**
      * @return string[]
      */
-    public function getRtrEvents()
+    public function getRtrEvents(): ?array
     {
         return $this->rtrEvents;
     }
+
     /**
      * @param string[] $rtrEvents
      *
      * @return self
      */
-    public function setRtrEvents(array $rtrEvents = null)
+    public function setRtrEvents(?array $rtrEvents): self
     {
         $this->rtrEvents = $rtrEvents;
+
         return $this;
     }
+
     /**
      * @return string[]
      */
-    public function getRtrKeywords()
+    public function getRtrKeywords(): ?array
     {
         return $this->rtrKeywords;
     }
+
     /**
      * @param string[] $rtrKeywords
      *
      * @return self
      */
-    public function setRtrKeywords(array $rtrKeywords = null)
+    public function setRtrKeywords(?array $rtrKeywords): self
     {
         $this->rtrKeywords = $rtrKeywords;
+
         return $this;
     }
+
     /**
      * @return string[]
      */
-    public function getRtrTags()
+    public function getRtrTags(): ?array
     {
         return $this->rtrTags;
     }
+
     /**
      * @param string[] $rtrTags
      *
      * @return self
      */
-    public function setRtrTags(array $rtrTags = null)
+    public function setRtrTags(?array $rtrTags): self
     {
         $this->rtrTags = $rtrTags;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -813,14 +936,16 @@ class RetrescoDocument
     {
         return $this->payload;
     }
+
     /**
      * @param mixed $payload
      *
      * @return self
      */
-    public function setPayload($payload = null)
+    public function setPayload($payload): self
     {
         $this->payload = $payload;
+
         return $this;
     }
 }
