@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class RetrescoEntityLinks
@@ -8,21 +16,24 @@ class RetrescoEntityLinks
      * @var string[]
      */
     protected $links;
+
     /**
      * @return string[]
      */
-    public function getLinks()
+    public function getLinks(): ?\ArrayObject
     {
         return $this->links;
     }
+
     /**
      * @param string[] $links
      *
      * @return self
      */
-    public function setLinks(\ArrayObject $links = null)
+    public function setLinks(?\ArrayObject $links): self
     {
         $this->links = $links;
+
         return $this;
     }
 }
