@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class RetrescoTopicPage
@@ -68,40 +76,47 @@ class RetrescoTopicPage
      * @var string
      */
     protected $url;
+
     /**
      * @return string
      */
-    public function getCanonicalTag()
+    public function getCanonicalTag(): ?string
     {
         return $this->canonicalTag;
     }
+
     /**
      * @param string $canonicalTag
      *
      * @return self
      */
-    public function setCanonicalTag($canonicalTag = null)
+    public function setCanonicalTag(?string $canonicalTag): self
     {
         $this->canonicalTag = $canonicalTag;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getDocId()
+    public function getDocId(): ?string
     {
         return $this->docId;
     }
+
     /**
      * @param string $docId
      *
      * @return self
      */
-    public function setDocId($docId = null)
+    public function setDocId(?string $docId): self
     {
         $this->docId = $docId;
+
         return $this;
     }
+
     /**
      * @return mixed
      */
@@ -109,235 +124,276 @@ class RetrescoTopicPage
     {
         return $this->itemsPerPage;
     }
+
     /**
      * @param mixed $itemsPerPage
      *
      * @return self
      */
-    public function setItemsPerPage($itemsPerPage = null)
+    public function setItemsPerPage($itemsPerPage): self
     {
         $this->itemsPerPage = $itemsPerPage;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getMetaDescription()
+    public function getMetaDescription(): ?string
     {
         return $this->metaDescription;
     }
+
     /**
      * @param string $metaDescription
      *
      * @return self
      */
-    public function setMetaDescription($metaDescription = null)
+    public function setMetaDescription(?string $metaDescription): self
     {
         $this->metaDescription = $metaDescription;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getMetaTitle()
+    public function getMetaTitle(): ?string
     {
         return $this->metaTitle;
     }
+
     /**
      * @param string $metaTitle
      *
      * @return self
      */
-    public function setMetaTitle($metaTitle = null)
+    public function setMetaTitle(?string $metaTitle): self
     {
         $this->metaTitle = $metaTitle;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getName()
+    public function getName(): ?string
     {
         return $this->name;
     }
+
     /**
      * @param string $name
      *
      * @return self
      */
-    public function setName($name = null)
+    public function setName(?string $name): self
     {
         $this->name = $name;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getQuery()
+    public function getQuery(): ?string
     {
         return $this->query;
     }
+
     /**
      * @param string $query
      *
      * @return self
      */
-    public function setQuery($query = null)
+    public function setQuery(?string $query): self
     {
         $this->query = $query;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getRedirect()
+    public function getRedirect(): ?string
     {
         return $this->redirect;
     }
+
     /**
      * @param string $redirect
      *
      * @return self
      */
-    public function setRedirect($redirect = null)
+    public function setRedirect(?string $redirect): self
     {
         $this->redirect = $redirect;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getSection()
+    public function getSection(): ?string
     {
         return $this->section;
     }
+
     /**
      * @param string $section
      *
      * @return self
      */
-    public function setSection($section = null)
+    public function setSection(?string $section): self
     {
         $this->section = $section;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getSuperTitle()
+    public function getSuperTitle(): ?string
     {
         return $this->superTitle;
     }
+
     /**
      * @param string $superTitle
      *
      * @return self
      */
-    public function setSuperTitle($superTitle = null)
+    public function setSuperTitle(?string $superTitle): self
     {
         $this->superTitle = $superTitle;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTeaser()
+    public function getTeaser(): ?string
     {
         return $this->teaser;
     }
+
     /**
      * @param string $teaser
      *
      * @return self
      */
-    public function setTeaser($teaser = null)
+    public function setTeaser(?string $teaser): self
     {
         $this->teaser = $teaser;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTeaserImg()
+    public function getTeaserImg(): ?string
     {
         return $this->teaserImg;
     }
+
     /**
      * @param string $teaserImg
      *
      * @return self
      */
-    public function setTeaserImg($teaserImg = null)
+    public function setTeaserImg(?string $teaserImg): self
     {
         $this->teaserImg = $teaserImg;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTeaserImgSubline()
+    public function getTeaserImgSubline(): ?string
     {
         return $this->teaserImgSubline;
     }
+
     /**
      * @param string $teaserImgSubline
      *
      * @return self
      */
-    public function setTeaserImgSubline($teaserImgSubline = null)
+    public function setTeaserImgSubline(?string $teaserImgSubline): self
     {
         $this->teaserImgSubline = $teaserImgSubline;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTitle()
+    public function getTitle(): ?string
     {
         return $this->title;
     }
+
     /**
      * @param string $title
      *
      * @return self
      */
-    public function setTitle($title = null)
+    public function setTitle(?string $title): self
     {
         $this->title = $title;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getTopicType()
+    public function getTopicType(): ?string
     {
         return $this->topicType;
     }
+
     /**
      * @param string $topicType
      *
      * @return self
      */
-    public function setTopicType($topicType = null)
+    public function setTopicType(?string $topicType): self
     {
         $this->topicType = $topicType;
+
         return $this;
     }
+
     /**
      * @return string
      */
-    public function getUrl()
+    public function getUrl(): ?string
     {
         return $this->url;
     }
+
     /**
      * @param string $url
      *
      * @return self
      */
-    public function setUrl($url = null)
+    public function setUrl(?string $url): self
     {
         $this->url = $url;
+
         return $this;
     }
 }
