@@ -1,5 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file has been auto generated by Jane,
+ *
+ * Do no edit it directly.
+ */
+
 namespace telekurier\RetrescoClient\Model;
 
 class RetrescoTopicPages
@@ -12,38 +20,44 @@ class RetrescoTopicPages
      * @var int
      */
     protected $numFound;
+
     /**
      * @return RetrescoTopicPage[]
      */
-    public function getDocs()
+    public function getDocs(): ?array
     {
         return $this->docs;
     }
+
     /**
      * @param RetrescoTopicPage[] $docs
      *
      * @return self
      */
-    public function setDocs(array $docs = null)
+    public function setDocs(?array $docs): self
     {
         $this->docs = $docs;
+
         return $this;
     }
+
     /**
      * @return int
      */
-    public function getNumFound()
+    public function getNumFound(): ?int
     {
         return $this->numFound;
     }
+
     /**
      * @param int $numFound
      *
      * @return self
      */
-    public function setNumFound($numFound = null)
+    public function setNumFound(?int $numFound): self
     {
         $this->numFound = $numFound;
+
         return $this;
     }
 }
